Keep Songchain Lens diagnostics from crashing
